+++ source/extensions/filters/network/well_known_names.h
@@ -26,6 +26,8 @@ public:
   const std::string MongoProxy = "envoy.filters.network.mongo_proxy";
   // MySQL proxy filter
   const std::string MySQLProxy = "envoy.filters.network.mysql_proxy";
+  // Proxy Protocol filter
+  const std::string ProxyProtocol = "envoy.filters.network.proxy_protocol";
   // Rate limit filter
   const std::string RateLimit = "envoy.filters.network.ratelimit";
   // Redis proxy filter